fix(format_value) handle unset value
Zelda was throwing an error when (I believe) a field was saved with no
value, because it was trying to format a value that didn't exist. This
is behavior we actually want to allow (i.e. the field is not required)
so this fix simply returns the content passed for formatting if the
`value` portion of the array is not set.